refactor: user route and directory to userProfile
